% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_basis.R
\name{change_basis}
\alias{change_basis}
\title{Performs a change-of-basis for bivariate functional data}
\usage{
change_basis(Y_list, rot_mat)
}
\arguments{
\item{Y_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{rot_mat}{Matrix, containing the coordinates of the linear transformation.
Typically a rotation matrix.}
}
\value{
List, containing the transformed sampling points and original
observations.
}
\description{
Given a rotational matrix, a change of basis is performed for the sampling
points of functional data in the form of bivariate sheets. Relevant for
anisotropic functional data with common design.
}
