% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_estim.R
\name{estimate_angle}
\alias{estimate_angle}
\title{Estimates the angle between two basis vectors up to a reflection}
\usage{
estimate_angle(X_list, xout, delta)
}
\arguments{
\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{xout}{Vector, containing the evaluation points at which
the angle should be computed.}

\item{delta}{Numeric, determining the spacings.}
}
\value{
List, containing:
\itemize{
\item \strong{$g_hat} Estimated function of angle.
\item \strong{$alpha_acot} Estimated angle by applying arccot.
\item \strong{$alpha_atan} Estimated angle by applying arctan.
\item \strong{$H_min} Estimated minimum regularity.
\item \strong{$sigma_hat} Estimated noise standard deviation.
}
}
\description{
Estimates the angle between two basis vectors up to a reflection
}
