% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Estimate the noise at the observed points of functional surfaces}
\usage{
estimate_sigma(X_list)
}
\arguments{
\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}
}
\value{
Matrix, where the number of rows and columns are each given by the
length of the sampling points.
}
\description{
Estimate the noise at the observed points of functional surfaces
}
