% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{fbm_sheet}
\alias{fbm_sheet}
\title{Simulates fractional brownian sheet}
\usage{
fbm_sheet(t_n, e_n, alpha, H1, H2)
}
\arguments{
\item{H1}{Numeric, indicating the Hölder exponent of the first fractional
brownian motion.}

\item{H2}{Numeric, indicating the Hölder exponent of the second fractional
brownian motion.}

\item{t1_n}{Numeric, containing the number of evaluation points for the
output along the direction of \code{u1}.}

\item{t2_n}{Numeric, containing the number of evaluation points for the
output along the direction of \code{u2}.}

\item{e1_n}{Numeric, containing the number of evaluation points along
the direction of the first canonical basis.}

\item{e2_n}{Numeric, containing the number of evaluation points along
the direction of the second canonical basis.}
}
\value{
Matrix, containing the fractional brownian sheets
observed on the canonical bases.
}
\description{
Simulates a fractional brownian sheet on the two-dimensional canonical basis
based on the product of two fractional brownian motions (fbms) on another basis.
The individual fbms are simulated using fast fourier transforms for increased
speed; see \code{fbm_fft}.
}
