% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{fbm_sheet_var}
\alias{fbm_sheet_var}
\title{Simulates fractional brownian sheet with time-varying anisotropy}
\usage{
fbm_sheet_var(t_n, e_n, alpha_fun, H1, H2, b_n = 10000)
}
\arguments{
\item{t_n}{Numeric, containing the number of evaluation points for the
output along the direction of anisotropic bases.}

\item{e_n}{Numeric, containing the number of evaluation points along
the direction of the canonical bases.}

\item{alpha_fun}{Function, indicating how to construct the time-varying angles.}

\item{H1}{Numeric, indicating the Hölder exponent of the first fractional
brownian motion.}

\item{H2}{Numeric, indicating the Hölder exponent of the second fractional
brownian motion.}

\item{b_n}{Numeric, indicating the number of evaluation points to simulate
the fBms on.}
}
\value{
Matrix, containing the fractional brownian sheets
observed on the canonical bases.
}
\description{
Simulates a fractional brownian sheet on the two-dimensional canonical basis
based on the sum of two fractional brownian motions (fbms) on another basis.
The individual fbms are simulated using fast fourier transforms for increased
speed; see \code{fbm_fft}. This function simulates brownian sheets with time-varying
angles; for a constant angle, please use \code{fbm_sheet} instead.
}
