% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{fbm_sum}
\alias{fbm_sum}
\title{Simulate isotropic process based on fractional brownian motions}
\usage{
fbm_sum(H1, H2, n, endpoint)
}
\description{
Simulate isotropic process based on fractional brownian motions
}
