% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g_compute}
\alias{g_compute}
\title{Computes the true estimating equations for the purposes of comparison.}
\usage{
g_compute(alpha, H1, H2, delta)
}
\arguments{
\item{alpha}{Numeric, the true angle.}

\item{H1}{Numeric, the regularity along the first basis.}

\item{H2}{Numeric, the regularity along the second basis.}

\item{delta}{Numeric, the spacings used in estimation.}
}
\value{
Numeric, the true tangent or cotangent of the angle.
}
\description{
Computes the true estimating equations for the purposes of comparison.
}
