% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{H_replace}
\alias{H_replace}
\title{Computation of regularity in the presence of noise}
\usage{
H_replace(theta_2delta, theta_delta, sigma)
}
\arguments{
\item{theta_2delta}{Numeric.}

\item{theta_delta}{Numeric.}

\item{sigma}{Numeric.}
}
\value{
Numeric.
}
\description{
Auxiliary function used in the computation of the regularity along some
directional vector.
}
