% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{H_sheets}
\alias{H_sheets}
\title{Computes the Hölder regularity of bivariate functional data}
\usage{
H_sheets(X_list, tout, delta)
}
\arguments{
\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{tout}{Matrix / dataframe with 2 dimensions, specifying all the
coordinates in the \code{x} and \code{y} axis.}

\item{delta}{Numeric, determining the spacings.}
}
\value{
List, containing the elements:
\itemize{
\item \strong{H} Matrix, containing the estimated regularity
on the dimensional grid \code{tout}.
\item \strong{theta_e2} Matrix, containing the thetas on the \code{e2} basis computed with
delta.
\item \strong{theta_e1} Matrix, containing the thetas on the \code{e1} basis computed with
delta.
-\strong{sigma_hat} Numeric, the estimated noise.
}
}
\description{
Computes the Hölder regularity of bivariate functional data
}
