% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_estim.R
\name{identify_angle}
\alias{identify_angle}
\title{Find the unique angle between the two basis vectors that provides the
maximising regularity}
\usage{
identify_angle(angles, X_list, dout, xout, sigma)
}
\arguments{
\item{angles}{List, for example the output obtained by \code{estimate_angle}, containing:
\itemize{
\item \strong{$g_hat} Numeric, estimated function of angle.
\item \strong{$alpha_acot} Numeric, estimated angle by applying arccot.
\item \strong{$alpha_atan} Numeric, estimated angle by applying arctan.
\item \strong{$H_min} Numeric, estimated minimum regularity.
\item \strong{$sigma_hat} Numeric, estimated noise standard deviation.
}}

\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{dout}{Vector, containing the grid of spacings (i.e delta) to compute
the regularity.}

\item{xout}{Vector, containing the evaluation points along one 1 dimension
to compute the regularity. The cartesian product is taken to produce the
2D grid.}

\item{sigma}{Numeric, the standard deviation of the noise,
for example outputted by \code{estimate_angle}.}
}
\value{
List, containing the identified angle and the associated estimated
regularity averaged over the grid of spacings.
}
\description{
Find the unique angle between the two basis vectors that provides the
maximising regularity
}
