% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{L_sheets}
\alias{L_sheets}
\title{Estimates the Hölder constant along some directional vector}
\usage{
L_sheets(X_list, xout, delta, v, Hv)
}
\arguments{
\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{xout}{Vector, containing the evaluation points at which
the estimates should be computed along one dimension. Computation is done
over the cartesian product of \code{xout} with itself.}

\item{delta}{Numeric, determining the spacings.}

\item{v}{Vector, determining the direction at which the Hölder constant should
be estimated.}

\item{Hv}{Numeric, containing the Hölder regularity along direction \code{v}.}
}
\value{
Numeric, containing the Hölder constant.
}
\description{
Computation is done using the mean-squared continuity that arises from the class
of stochastic process considered. Estimates are first computed on a grid of time
points, before the average over the time points are taken to stabilise estimates.
}
