% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{theta_sheets}
\alias{theta_sheets}
\title{Computes the mean squared deviations of a surface on a grid on points}
\usage{
theta_sheets(X_list, tout, delta, e)
}
\arguments{
\item{X_list}{List, containing the following elements:
-\strong{$t} Vector of sampling points,
-\strong{X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.}

\item{tout}{Data frame of points on which the thetas should be computed,
resulting from \code{expand.grid}.}

\item{delta}{Numeric, determining the spacings.}

\item{e}{List, containing the coordinates of the basis vectors.}
}
\value{
Matrix, containing the mean squared deviations on the evaluation
points \code{tout}.
}
\description{
Computes the mean squared deviations of a surface on a grid on points
}
