% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_estim.R
\name{thresh_angle}
\alias{thresh_angle}
\title{Determines if process is anisotropic based on thresholding}
\usage{
thresh_angle(H_min, H_max, tau)
}
\arguments{
\item{H_min}{Numeric, indicating the estimated minimum regularity.}

\item{H_max}{Numeric, indicating the estimated maximum regularity.}

\item{tau}{Numeric, indicating the thresholding constant.}
}
\value{
Numeric, with TRUE indicating anisotropy and FALSE otherwise.
}
\description{
Determines if process is anisotropic based on thresholding
}
