% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_estim.R
\name{thresh_tau}
\alias{thresh_tau}
\title{Computes the thresholding parameter used for anisotropic detection}
\usage{
thresh_tau(n_dout, delta, M0, N, H_min, c_tau)
}
\arguments{
\item{n_dout}{Numeric, number of grid points in the spacing used for
identification.}

\item{delta}{Numeric, indicating the spacing.}

\item{M0}{Numeric, number of observed points along each surface.}

\item{N}{Numeric, number of surfaces.}

\item{H_min}{Numeric, the estimated minimum regularity along the dimensions.}

\item{c_tau}{Numeric, a constant which accompanies the rate.}
}
\value{
Numeric, indicating the thresholding constant.
}
\description{
Computes the thresholding parameter used for anisotropic detection
}
