% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_basis.R
\name{angle_correct}
\alias{angle_correct}
\title{Compute the correction for estimating equation of angle}
\usage{
angle_correct(g_hat, alpha_hat, delta, Hmax, Hmin)
}
\arguments{
\item{g_hat}{Numeric, the preliminary estimate of the tangent or cotangent of the angle.}

\item{alpha_hat}{Numeric, the preliminary estimate of the identified angle.}

\item{delta}{Numeric, the spacing used for estimation.}

\item{Hmax}{Numeric, the maximum regularity.}

\item{Hmin}{Numeric, the minimum regularity.}
}
\value{
List, containing the corrected angle, corrected function of angle and
the correction term.
}
\description{
Correction is based on the remainder term that arises from the estimating
equation used for the directional regularity. Should only be used once to
avoid introducing additional dependence between estimates.
}
