% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_detect.R
\name{ani_detect}
\alias{ani_detect}
\title{Performs anisotropic detection by thresholding}
\usage{
ani_detect(
  X_list,
  beta_n,
  g_adj,
  alpha_adj,
  H_min,
  H_max,
  delta_grid,
  sigma,
  H_pow = 1/3
)
}
\arguments{
\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{beta_n}{Numeric, number of random angles to take.}

\item{g_adj}{Numeric, the estimated and corrected function of angle.}

\item{alpha_adj}{Numeric, the estimated and correction angle.}

\item{H_min}{Numeric, the estimated minimum regularity.}

\item{H_max}{Numeric, the estimated maximum regularity.}

\item{delta_grid}{Vector, the grid of spacings used for estimation.}

\item{sigma}{Numeric, the estimated noise.}

\item{H_pow}{Numeric, the power of the rate term.}
}
\value{
List, containing the following elements:
\itemize{
\item \strong{$indic}: Numeric, with 1 indicator anisotropy and 0 otherwise.
\item \strong{$tau} : Numeric, the threshold value used to test for anisotropy.
}
}
\description{
Performs anisotropic detection by thresholding
}
