% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_estim.R
\name{estimate_angle}
\alias{estimate_angle}
\title{Estimates the angle between two basis vectors up to a reflection}
\usage{
estimate_angle(X_list, xout, delta)
}
\arguments{
\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{xout}{Vector, containing the evaluation points at which
the angle should be computed.}

\item{delta}{Numeric, determining the spacings.}
}
\value{
Vector, containing the angles identified by arccot and arctan.
}
\description{
Estimates the angle between two basis vectors up to a reflection
}
