% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{fbm_sheet}
\alias{fbm_sheet}
\title{Simulates fractional brownian sheet with constant anisotropy}
\usage{
fbm_sheet(t_n, e_n, alpha, H1, H2)
}
\arguments{
\item{t_n}{Numeric, the number of evaluation points for the
output along the direction of the non-canonical bases.}

\item{e_n}{Numeric, the number of evaluation points along
the direction of the canonical bases}

\item{alpha}{Numeric, the angle between the first non-canonical and canonical
basis. Must be a number between \verb{[0, pi/2]}.}

\item{H1}{Numeric, the Hölder exponent of the first fractional
brownian motion.}

\item{H2}{Numeric, the Hölder exponent of the second fractional
brownian motion.}
}
\value{
Matrix, containing the fractional brownian sheets
observed on the canonical bases.
}
\description{
Simulates a fractional brownian sheet on the two-dimensional canonical basis
based on the product of two fractional brownian motions (fbms) on another basis.
The individual fbms are simulated using fast fourier transforms for increased
speed; see \code{fbm_fft}.
}
