% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{H_sheets_dir}
\alias{H_sheets_dir}
\title{Compute the Hölder regularity along the direction of input bases}
\usage{
H_sheets_dir(X_list, tout, delta, base_list, sigma)
}
\arguments{
\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{tout}{Matrix / dataframe with 2 dimensions, specifying all the
coordinates in the \code{x} and \code{y} axis.}

\item{delta}{Numeric, determining the spacings.}

\item{base_list}{List, with each element containing the coordinates of
each vector.}

\item{sigma}{Numeric, indicating the noise level, for example from the
\code{estimated_sigma} function.}
}
\value{
Vector, containing the estimated regularity along each direction
of \code{base_list}. The order of outputs corresponds to the order of inputs in
\code{base_list}.
}
\description{
Compute the Hölder regularity along the direction of input bases
}
