% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_basis.R
\name{identify_angle}
\alias{identify_angle}
\title{Find the unique angle between the two basis vectors that provides the
maximising regularity}
\usage{
identify_angle(angles, X_list, dout, xout)
}
\arguments{
\item{angles}{Vector, containing the arccot and arctan of the angle, for
example outputted by the function \code{estimate_angle}.}

\item{X_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points,
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{dout}{Vector, containing the grid of spacings (i.e delta) to compute
the regularity.}

\item{xout}{Vector, containing the evaluation points along one 1 dimension
to compute the regularity. The cartesian product is taken to produce the
2D grid.}
}
\value{
Numeric, containing the identified angle.
}
\description{
Find the unique angle between the two basis vectors that provides the
maximising regularity
}
