% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing.R
\name{ksmooth_bi}
\alias{ksmooth_bi}
\title{Bi-variate kernel smoothing for functional data on common design}
\usage{
ksmooth_bi(Y_list, bw_vec, xout)
}
\arguments{
\item{Y_list}{List, containing the following elements:
\itemize{
\item \strong{$t} Vector of sampling points.
\item \strong{$X} Matrix of observed points, measured on the bi-dimensional grid containing
cartesian product of \verb{$t} with itself.
}}

\item{bw_vec}{Vector of bandwidths, corresponding to the diagonal elements of
the bandwidth matrix.}

\item{xout}{Vector of evaluation points along one dimension. Smoothing is performed
on the cartesian product of \code{xout} with itself.}
}
\value{
List, corresponding to the smooth curves.
}
\description{
Bi-variate kernel smoothing for functional data on common design
}
